pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "political-monitoring-agent"
version = "1.0.0"
description = "Political document monitoring agent for Kodosumi on Ray"
authors = [{name = "Political Monitoring Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Kodosumi Framework
    "kodosumi==0.9.2",
    # Core Framework
    "ray[serve]==2.46.0",
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-multipart==0.0.6",
    # Workflow & State Management
    "langgraph==0.2.3",
    "pydantic>=2.7.4,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    # Document Processing
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "markdown==3.5.1",
    "beautifulsoup4==4.12.2",
    "lxml==4.9.3",
    # AI/ML Components
    "numpy==1.26.2",
    "scikit-learn==1.3.2",
    "tiktoken>=0.7,<1",
    # LLM Integration
    "langchain==0.2.16",
    "langchain-openai==0.1.25",
    "langchain-anthropic==0.1.23",
    "langchain-community==0.2.16",
    "langfuse==2.36.1",
    # Output Generation
    "jinja2==3.1.2",
    "tabulate==0.9.0",
    # Utilities
    "python-dotenv==1.0.0",
    "pyyaml==6.0.1",
    "aiofiles==23.2.1",
    "httpx==0.25.2",
    # Monitoring & Logging
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    "azure-storage-blob>=12.25.1",
    "azure-identity>=1.23.0",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",
    "pytest-cov==4.1.0",
    "faker==20.1.0",
]
dev = [
    "ruff==0.1.7",
    "mypy==1.7.0",
    "pre-commit==3.5.0",
]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",
    "pytest-cov==4.1.0",
    "faker==20.1.0",
    "ruff==0.1.7",
    "mypy==1.7.0",
    "pre-commit==3.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "S", "T20"]
ignore = ["E501", "S101"]  # Line too long, assert statements

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=90",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
